enable logger and callbacks
